Add more exclusions to `override_type` method:
* Bail if no post type change
* Bail if post ID does not match posted ID
* Check both required caps before performing other logic

This is an addition to several other exclusions, all of which are an attempt to be 100% certain that the post type being changed is only for the requested post ID, and not any other related post.

This gets tricky during nested or recursive calls, so adding the above bit of confirmation on the ID should help alleviate most of those issues.

See #9.